fix customerDto and modified remaining long Id into Int

+++ backend/src/main/java/com/pants/backend/controller/ReceiptController.java
@@ -34,7 +34,7 @@ public class ReceiptController {
     }
 
     @PostMapping("/{reservationId}")
-    public Receipt createReceipt(@PathVariable Long reservationId) {
+    public Receipt createReceipt(@PathVariable Integer reservationId) {
 
         Reservation reservation = reservationRepository.findById(reservationId).orElse(null);
